+++ src/Resources/contao/dca/tl_contao_speisekarte_speisen.php
@@ -8,6 +8,15 @@
  *  note       : 20210806°0911 Start implementing image
  */
 
+use Contao\Backend;
+use Contao\BackendUser;
+use Contao\Config;
+use Contao\CoreBundle\Security\ContaoCorePermissions;
+use Contao\Image;
+use Contao\StringUtil;
+use Contao\System;
+use LinkingYou\ContaoSpeisekarte\Model\ContaoSpeisekarteSpeisenModel;
+
 $GLOBALS['TL_DCA']['tl_contao_speisekarte_speisen'] = array
 (
 
@@ -21,7 +30,8 @@ $GLOBALS['TL_DCA']['tl_contao_speisekarte_speisen'] = array
         (
             'keys' => array
             (
-                'id' => 'primary'
+                'id' => 'primary',
+                'pid,published' => 'index'
             )
         )
     ),
@@ -35,7 +45,7 @@ $GLOBALS['TL_DCA']['tl_contao_speisekarte_speisen'] = array
             'fields'                  => array('sorting'),
             'flag'                    => 11,
             'panelLayout'             => 'search,limit',
-            'child_record_callback'   => array('Speisen', 'getSpeisen'),
+            'child_record_callback'   => array('tl_contao_speisekarte_speisen', 'getSpeisen'),
             'headerFields'            => array('titel')
         ),
         'label' => array
@@ -73,6 +83,13 @@ $GLOBALS['TL_DCA']['tl_contao_speisekarte_speisen'] = array
                 'icon'                => 'delete.gif',
                 'attributes'          => 'onclick="if(!confirm(\'' . $GLOBALS['TL_LANG']['MSC']['deleteConfirm'] . '\'))return false;Backend.getScrollOffset()"'
             ),
+            'toggle' => array
+            (
+              'href'                => 'act=toggle&amp;field=published',
+              'icon'                => 'visible.svg',
+              'button_callback'     => array('tl_contao_speisekarte_speisen', 'toggleIcon')
+            ),
+
             'show' => array
             (
                 'label'               => &$GLOBALS['TL_LANG']['tl_contao_speisekarte_speisen']['show'],
@@ -98,8 +115,7 @@ $GLOBALS['TL_DCA']['tl_contao_speisekarte_speisen'] = array
     'palettes' => array
     (
         '__selector__'                => array(''),
-        ////'default'                 => '{titel_legend},nummer,titel,beschreibung;{Bild},picture;{Preise},menge,preis,menge2,preis2,menge3,preis3,einheit,grundpreis;{zusatzstoffe_legende},zusatzstoffe,allergene;'
-        'default'                     => '{titel_legend},nummer,titel,beschreibung;{Bild},dishpic;{Preise},menge,preis,menge2,preis2,menge3,preis3,einheit,grundpreis;{zusatzstoffe_legende},zusatzstoffe,allergene;'
+        'default'                     => '{titel_legend},nummer,titel,beschreibung;{dishpic_legend},dishpic;{preise_legend},menge,preis,menge2,preis2,menge3,preis3,einheit,grundpreis;{zusatzstoffe_legende},zusatzstoffe,allergene;{publish_legend},published;'
     ),
 
     // Subpalettes
@@ -146,6 +162,11 @@ $GLOBALS['TL_DCA']['tl_contao_speisekarte_speisen'] = array
             'inputType'               => 'text',
             'eval'                    => array(
                 'mandatory' => true,
+                 'allowHtml'  => true, 
+//                  'rte'        => 'tinyMCE' , 
+//                  'preserveTags' => false,
+//                  'rows'       => 1,  
+//                  'cols'       => 50, 
                 'maxlength' => 255,
                 'tl_class'  => 'w50 widget'
             ),
@@ -169,7 +190,7 @@ $GLOBALS['TL_DCA']['tl_contao_speisekarte_speisen'] = array
                 'rows'       => 4,                                     // added 20210805°092217
                 'rte'        => 'tinyMCE'                              // added 20210805°092218 'rich text editor'
              ),
-            'sql'                     => "varchar(10000) NOT NULL default ''"  // added 20210805°092119
+            'sql'                     => "text NULL"
         ),
 
  /*
@@ -214,7 +235,7 @@ $GLOBALS['TL_DCA']['tl_contao_speisekarte_speisen'] = array
                 'filesOnly'           => true,                                  // Avoid selecting a folder
                 'extensions'          => \Config::get('validImageTypes'),       // Limit file tree to certain file types
                 'tl_class'            => 'clr',                                 // Add the given CSS class(es) to the generated HTML
-                'mandatory'           => true                                   // If true the field cannot be empty
+                'mandatory'           => false                                   // If true the field cannot be empty
 
                 ////'fieldType'       => 'radio',                               // Select only one, not multiple as with checkbox
                 ////'files'           => true                                   // after outdated https://docs.contao.org/books/cookbook/en/custom-module/part2.html
@@ -314,26 +335,17 @@ $GLOBALS['TL_DCA']['tl_contao_speisekarte_speisen'] = array
             'label'                   => &$GLOBALS['TL_LANG']['tl_contao_speisekarte_speisen']['einheit'],
             'exclude'                 => true,
             'inputType'               => 'select',
-            'options'                 => array (
-                                            'Liter',
-                                            'g',       // possibly superfluous
-                                            'kg',      // possibly superfluous
-                                            'Stk'
-            ),
-            'eval'                    => array (
-                                             'includeBlankOption' => true,
-                                             'tl_class'           => 'w50 widget'
-                                         ),
+            'options'                 => array ('Liter', 'g', 'kg', 'Stk'),
+            'reference'               => &$GLOBALS['TL_LANG']['MSC']['contao_speisekarte']['einheit'],
+            'eval'                    => array ('includeBlankOption' => true, 'tl_class' => 'w50 widget'),
             'sql'                     => "varchar(255) NOT NULL default ''"
         ),
         'grundpreis' => array(
             'label'                   => &$GLOBALS['TL_LANG']['tl_contao_speisekarte_speisen']['grundpreis'],
             'exclude'                 => true,
             'inputType'               => 'checkbox',
-            'eval' => array(
-                'tl_class' => 'w50 widget'
-            ),
-            'sql'                     => "varchar(10) NOT NULL default ''"
+            'eval'                    => array('tl_class' => 'w50 widget'),
+            'sql'                     => "char(1) NOT NULL default ''"
         ),
         'zusatzstoffe' => array(
             'label'                   => &$GLOBALS['TL_LANG']['tl_contao_speisekarte_speisen']['zusatzstoffe'],
@@ -362,12 +374,77 @@ $GLOBALS['TL_DCA']['tl_contao_speisekarte_speisen'] = array
             ),
             //'sql'                   => "varchar(255) NOT NULL default ''"
             'sql'                     => "text NULL"
+        ),
+        'published' => array
+        (
+          'exclude'                 => true,
+          'default'                 => true,
+          'toggle'                  => true,
+          'filter'                  => true,
+          'inputType'               => 'checkbox',
+          'eval'                    => array('doNotCopy'=>true),
+          'sql'                     => "char(1) NOT NULL default '1'"
         )
     )
 );
 
-class Speisen {
+class tl_contao_speisekarte_speisen extends Backend
+{
+    /**
+     * Import the back end user object
+     */
+    public function __construct()
+    {
+      parent::__construct();
+      $this->import(BackendUser::class, 'User');
+    }
+    
     function getSpeisen($a) {
         return $a['titel'] . '<br />' . $a['beschreibung'];
     }
+    
+    /**
+     * Return the "toggle visibility" button
+     *
+     * @param array  $row
+     * @param string $href
+     * @param string $label
+     * @param string $title
+     * @param string $icon
+     * @param string $attributes
+     *
+     * @return string
+     */
+    public function toggleIcon($row, $href, $label, $title, $icon, $attributes)
+    {
+      $security = System::getContainer()->get('security.helper');
+
+      // Check permissions AFTER checking the tid, so hacking attempts are logged
+      if (!$security->isGranted(ContaoCorePermissions::USER_CAN_EDIT_FIELD_OF_TABLE, 'tl_contao_speisekarte_speisen::published'))
+      {
+        return '';
+      }
+
+      $href .= '&amp;id=' . $row['id'];
+
+      if (!$row['published'])
+      {
+        $icon = 'invisible.svg';
+      }
+
+      $objSpeise = ContaoSpeisekarteSpeisenModel::findById($row['id']);
+
+      if (!$security->isGranted(ContaoCorePermissions::USER_CAN_EDIT_ARTICLES, $objSpeise->row()))
+      {
+        if ($row['published'])
+        {
+          $icon = preg_replace('/\.svg$/i', '_.svg', $icon);
+        }
+
+        return Image::getHtml($icon) . ' ';
+      }
+
+      return '<a href="' . $this->addToUrl($href) . '" title="' . StringUtil::specialchars($title) . '" onclick="Backend.getScrollOffset();return AjaxRequest.toggleField(this,true)">' . Image::getHtml($icon, $label, 'data-icon="' . Image::getPath('visible.svg') . '" data-icon-disabled="' . Image::getPath('invisible.svg') . '" data-state="' . ($row['published'] ? 1 : 0) . '"') . '</a> ';
+    }
+
 }